+++ plugins/builtin/include/content/helpers/diagrams.hpp
@@ -18,6 +18,7 @@
 
 #include <atomic>
 #include <random>
+#include <hex/helpers/auto_reset.hpp>
 
 namespace hex {
 
@@ -55,7 +56,7 @@ namespace hex {
 
                 std::map<u64, std::vector<u8>> orderedData;
                 for (u32 i = 0; i < sequenceCount; i++) {
-                    ssize_t offset = random() % size;
+                    i64 offset = random() % size;
 
                     std::vector<u8> sequence;
                     sequence.resize(std::min<size_t>(sequenceCount, size - offset));
@@ -92,7 +93,7 @@ namespace hex {
 
                 std::map<u64, std::vector<u8>> orderedData;
                 for (u32 i = 0; i < sequenceCount; i++) {
-                    ssize_t offset = random() % inputBuffer.size();
+                    i64 offset = random() % inputBuffer.size();
 
                     std::vector<u8> sequence;
                     sequence.reserve(sampleSize);
